mouse look tinkering
Attempting to modify code to have the "tab" button make the cursor appear and turn off the mouse look, and then when hitting tab again, will make the cursor disappear and the mouse look turned on. I must read up on how to make a function and signal it to another .gd file.