Props (#6)
* mouse look tinkering

Attempting to modify code to have the "tab" button make the cursor appear and turn off the mouse look, and then when hitting tab again, will make the cursor disappear and the mouse look turned on. I must read up on how to make a function and signal it to another .gd file.

* mouselook on and off

still not working...

* mouselook inventory progress

Managed to make the mouselook turn on and off. Has something to do with:
"inventory_interface.show()" in the main .gd script. Investigate further.

* mouselook on and off switch

it now works. I'm going to work on the hotel now. I will add more features and items as I go along, while starting to build the interior of the hotel.

* last minute changes

* ignoreidea folder created